refactor(constructor): change file creation / naming
Revise 'prefix' argument in the class constructor to a file name. Remove '.settings' from the file name if one is provided, or use it as the file name if not.

The .json extension is also now ignored if passed to the constructor.

BREAKING CHANGE: file names are no longer including '.settings' if an argument is provided to the constructor.